Merge requirements files
The Vagrant box is now setup to only run the postgres server, and its
port forward to the host. Django should be ran from the host as well
and connect to postgres as a remote server. Due to these changes it's
no longer necessary to differentiate between local/deploy requirements.